+++ elements/layercontrol/src/helpers/layer-tools.js
@@ -1,9 +1,11 @@
 import { html, nothing } from "lit";
 
 /**
+ * Filters actions based on provided tools and layer properties.
  *
- * @param {Array<string>} tools
- * @return {Array<string>}
+ * @param {Array<string>} tools - Array of tool names.
+ * @param {import("ol/layer").Layer} layer - The native OL layer
+ * @returns {Array<string>} - Filtered array of action names.
  */
 export const _parseActions = (tools, layer) => {
   return tools?.filter((t) =>
@@ -14,36 +16,47 @@ export const _parseActions = (tools, layer) => {
 };
 
 /**
+ * Parses tools based on provided tools and layer properties.
  *
- * @param {Array<string>} tools
+ * @param {Array<string>} tools - Array of tool names.
+ * @param {import("ol/layer").Layer} layer - The native OL layer
+ * @returns {Array<string>} - Parsed array of tool names.
  */
 export const _parseTools = (tools, layer) =>
   tools?.filter((t) => {
     let pass = true;
-    if (["remove", "sort"].includes(t)) {
-      pass = false;
-    }
-    if (t === "info") {
-      pass = layer.get("description");
-    }
-    if (t === "config") {
-      pass = layer.get("layerConfig");
-    }
+    if (["remove", "sort"].includes(t)) pass = false;
+    if (t === "info") pass = layer.get("description");
+    if (t === "config") pass = layer.get("layerConfig");
+
     return pass;
   });
 
+/**
+ * Generates a button element based on the provided tool and style preference.
+ *
+ * @param {string} tool - Tool name for the button.
+ * @param {boolean} unstyled - Style preference for the button.
+ * @returns {import("lit").HTMLTemplateResult} - Button element.
+ */
 export const Button = (tool, unstyled) => html`
   <button slot="${tool}-icon" class="icon">${unstyled ? tool : nothing}</button>
 `;
 
-export const removeButton = (EOxLayerControlTools) => html`
+/**
+ * Generates a remove button element with event handling and layer control logic.
+ *
+ * @param {import("../components/layer-tools").EOxLayerControlLayerTools} EOxLayerControlLayerTools - Instance of EOxLayerControlLayerTools
+ * @returns {import("lit").HTMLTemplateResult} - Remove button element.
+ */
+export const removeButton = (EOxLayerControlLayerTools) => html`
   <button
     class="remove-icon icon"
     @click=${() => {
-      const { layer } = EOxLayerControlTools;
+      const { layer } = EOxLayerControlLayerTools;
       layer?.set("layerControlOptional", true);
       layer?.setVisible(false);
-      EOxLayerControlTools.dispatchEvent(
+      EOxLayerControlLayerTools.dispatchEvent(
         new CustomEvent("changed", {
           detail: layer,
           bubbles: true,
@@ -51,10 +64,16 @@ export const removeButton = (EOxLayerControlTools) => html`
       );
     }}
   >
-    ${EOxLayerControlTools.unstyled ? "x" : nothing}
+    ${EOxLayerControlLayerTools.unstyled ? "x" : nothing}
   </button>
 `;
 
+/**
+ * Generates a sort button element based on the style preference.
+ *
+ * @param {boolean} unstyled - Style preference for the button.
+ * @returns {import("lit").HTMLTemplateResult} - Sort button element.
+ */
 export const sortButton = (unstyled) => html`
   <button class="sort-icon icon drag-handle">
     ${unstyled ? "sort" : nothing}
