+++ elements/layercontrol/src/methods/add-layer/open-close-tab.js
@@ -3,7 +3,7 @@
  * Updates relevant properties based on the selected tab.
  *
  * @param {string} tab - The tab identifier to open or close.
- * @param {import("../../components/addLayers").EOxLayerControlAddLayers} EoxLayerControlAddLayers - Instance of EOxLayerControlAddLayers.
+ * @param {import("../../components/add-layers").EOxLayerControlAddLayers} EoxLayerControlAddLayers - Instance of EOxLayerControlAddLayers.
  */
 const openCloseTab = (tab, EoxLayerControlAddLayers) => {
   // Update the 'open' property of EOxLayerControlAddLayers based on the provided tab
